Move fuzz tests to ./test/fuzz directory
