+++ test/fuzz/index.test.cjs
@@ -8,7 +8,7 @@ const cp = require("child_process");
 const os = require("os");
 const process = require("process");
 
-const shescape = require("../index.cjs");
+const shescape = require("../../index.cjs");
 
 require("dotenv").config();
